requirements.txt
absl-py==2.0.0
acvl-utils==0.2.5
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
batchgenerators==0.25.1
beautifulsoup4==4.12.2
bleach==6.1.0
blosc2==3.0.0b4
cachetools==5.3.2
click==8.1.8
comm==0.2.1
connected-components-3d==3.23.0
contourpy==1.2.0
ConvKAN3D==0.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
einops==0.8.1
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
fonttools==4.47.0
fqdn==1.5.1
fsspec==2023.12.2
future==1.0.0
google-auth==2.26.1
google-auth-oauthlib==1.2.0
grpcio==1.60.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
imageio==2.37.0
importlib_resources==6.5.2
ipykernel==6.28.0
ipython==8.20.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.2
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kagglehub==0.3.11
kiwisolver==1.4.5
lazy_loader==0.4
linecache2==1.0.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdurl==0.1.2
MedPy==0.5.2
mistune==3.0.2
monai==1.4.0
mpmath==1.3.0
msgpack==1.1.0
nbclient==0.9.0
nbconvert==7.14.0
nbformat==5.9.2
ndindex==1.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nibabel==5.3.2
notebook_shim==0.2.3
numexpr==2.10.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
overrides==7.4.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pykan==0.2.8
pyparsing==3.1.1
pyprof==1.0.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2025.2
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.16.2
rsa==4.9
scikit-image==0.25.2
scikit-learn==1.4.2
scipy==1.15.2
Send2Trash==1.8.2
shellingham==1.5.4
SimpleITK==2.4.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
supervisor==4.2.5
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
termcolor==3.0.1
terminado==0.18.0
threadpoolctl==3.6.0
tifffile==2025.3.30
tinycss2==1.2.1
tomli==2.0.1
torch==2.1.0
torchio==0.20.8
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
traceback2==1.4.0
traitlets==5.14.1
triton==2.3.0
typer==0.15.3
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2025.2
unfoldNd==0.2.3
unittest2==1.1.0
uri-template==1.3.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.17.2
